migrate: move files to `packages/web` - auto update imports - creates minimal package.json and tsconfig - does not compile
