+++ packages/web/client/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@mergenb/web-client",
+  "version": "0.0.0",
+  "private": true,
+  "type": "commonjs"
+}
