+++ packages/web/client/src/components/MergeRow.tsx
@@ -11,9 +11,9 @@
 
 import React, { useEffect, useState, useMemo } from 'react';
 import CodeMirror, { Extension } from '@uiw/react-codemirror';
-import type { MergeRow as MergeRowType, ResolutionChoice } from './types';
+import type { MergeRow as MergeRowType, ResolutionChoice } from '../types';
 import { CellContent, mergeNBEditorStructure } from './CellContent';
-import { normalizeCellSource, selectNonConflictMergedCell } from '../../../packages/core/src/notebookUtils';
+import { normalizeCellSource, selectNonConflictMergedCell } from '../../../../core/src/notebookUtils';
 import { githubDark, githubLight } from '@uiw/codemirror-theme-github';
 
 /** Resolution state for a cell */
