+++ packages/web/client/src/hooks/useWebSocket.ts
@@ -4,8 +4,8 @@
  */
 
 import { useState, useEffect, useCallback, useRef } from 'react';
-import type { UnifiedConflictData, WSMessage } from './types';
-import * as logger from '../../../packages/core/src/logger';
+import type { UnifiedConflictData, WSMessage } from '../types';
+import * as logger from '../../../../core/src/logger';
 
 interface UseWebSocketResult {
     connected: boolean;
