+++ packages/web/client/src/store/resolverStore.ts
@@ -1,9 +1,9 @@
 import { enableMapSet } from 'immer';
 import { createStore, type StoreApi } from 'zustand/vanilla';
 import { immer } from 'zustand/middleware/immer';
-import { normalizeCellSource } from '../../../packages/core/src/notebookUtils';
-import { sortByPosition } from '../../../packages/core/src/positionUtils';
-import type { MergeRow as MergeRowType, NotebookCell, ResolutionChoice } from './types';
+import { normalizeCellSource } from '../../../../core/src/notebookUtils';
+import { sortByPosition } from '../../../../core/src/positionUtils';
+import type { MergeRow as MergeRowType, NotebookCell, ResolutionChoice } from '../types';
 
 enableMapSet();
