+++ packages/web/client/src/utils/reorderUtils.ts
@@ -1,4 +1,4 @@
-import type { MergeRow as MergeRowType } from './types';
+import type { MergeRow as MergeRowType } from '../types';
 
 interface IndexedRow {
     row: MergeRowType;
