+++ packages/web/client/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "../../..",
+    "noEmit": true
+  },
+  "include": [
+    "src/**/*"
+  ]
+}
