+++ packages/web/server/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@mergenb/web-server",
+  "version": "0.0.0",
+  "private": true,
+  "type": "commonjs"
+}
