+++ src/web/WebConflictPanel.ts
@@ -12,8 +12,8 @@
 import path from 'path';
 import * as vscode from 'vscode';
 import * as logger from '../../packages/core/src/logger';
-import { getWebServer } from './webServer';
-import { UnifiedConflict, UnifiedResolution, ResolvedRow, type WebConflictData } from './webTypes';
+import { getWebServer } from '../../packages/web/server/src/webServer';
+import { UnifiedConflict, UnifiedResolution, ResolvedRow, type WebConflictData } from '../../packages/web/server/src/webTypes';
 
 /**
  * Web-based panel for resolving notebook conflicts in the browser.
